requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=6.0.0
openpyxl>=3.1.0
pillow>=10.0.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
groq>=0.8.0
pdf2image>=1.17.0
pymupdf>=1.23.0
